bug ---order监听事件,计算总价抽出,orderItem 逻辑抽出
