+++ src/Components/OrderBox/orderBox.scss
@@ -6,7 +6,6 @@
   height : 390px;
   padding : 10px 20px;
   margin-bottom : 30px;
-  // margin-bottom : 30px;
   overflow : hidden;
   background-color : $white;
   border : 1px solid $line-grey;
